Moved CBinaryAsyncStream to bas.h and bas.cpp
