+++ include/bis.h
@@ -1,16 +1,10 @@
 #pragma once
+#include <bas.h>
 #include <prog.h>
 #include <iostream>
 #include <stdint.h>
 
 typedef unsigned char BYTE;
-
-enum class BASK : int
-{
-    Cd = 1,
-    Host = 0,
-    Max = 2
-};
 
 class CBinaryInputStream
 {
@@ -58,24 +52,4 @@ class CBinaryInputStream
     int32_t S32Read();
     float F32Read();
     void ReadStringSw(char** pachz);
-};
-
-class CBinaryAsyncStream {
-public:
-    BASK m_bask;
-    int m_fd;
-    uint32_t m_isector;
-    int m_cbSpooling;
-    BYTE* m_pbSpooling;
-    BYTE* m_abSpool;
-    int m_cb;
-    BYTE* m_pb;
-    int m_ibCur;
-    int m_cbUnspooled;
-    int m_cbFile;
-
-    CBinaryAsyncStream(void* pvSpool); // constructor
-    ~CBinaryAsyncStream(); // todo
-
-    void Close(); // todo
 };
