+++ src/CMakeLists.txt
@@ -1,6 +1,7 @@
 
 
 set (FILES
+	bas.cpp
 	bis.cpp
 	cat.cpp
 	clock.cpp
