+++ src/bis.cpp
@@ -1,7 +1,5 @@
 #include <bis.h>
 
-/* CBinaryInputStream class methods */
-
 uint32_t CBinaryInputStream::FOpenSector(uint32_t isector, uint32_t cb)
 {
     int iVar1;
@@ -261,21 +259,4 @@ void CBinaryInputStream::ReadStringSw(char** pachz)
     //Read(string_count, dst1);
     //dst1[string_count] = '\0';
     //*pachz = dst1;
-}
-
-/* CBinaryAsyncStream class methods */
-
-/* Constructor */
-CBinaryAsyncStream::CBinaryAsyncStream(void* pvSpool)
-{
-    this->m_isector = 0;
-    this->m_abSpool = (BYTE*)((int)pvSpool + 0x3fU & 0xffffffc0);
-    this->m_fd = -1;
-    this->m_pbSpooling = NULL;
-    this->m_pb = NULL;
-    this->m_cbFile = 0;
-    this->m_cbUnspooled = 0;
-    this->m_cbSpooling = 0;
-    this->m_ibCur = 0;
-    this->m_cb = 0;
 }
